Major overhauling of the room shifting code
I've decided to change a lot of the code of the room shifting. Mainly, I've decided to delegate finding the rooms and which direction they have to move, to the player script. This means that the rooms only need to be called upon by a function from player to move, which is what I've started to implement here.